# perf(resize): faster resize operation