No idea how often we have to resize the images, but we can improve the performance of the current resize by just initializing the buffer to the appropriate size and copying rows of pixels instead of px by px.

Quick benchmark
```
px copy 920x500 -> 1980x1080 x 139 ops/sec ±0.27% (81 runs sampled)
row copy 920x500 -> 1980x1080 x 839 ops/sec ±10.21% (44 runs sampled)
px copy 4x3 -> 1980x1080 x 159 ops/sec ±0.12% (83 runs sampled)
row copy 4x3 -> 1980x1080 x 2,049 ops/sec ±17.84% (36 runs sampled)
```

I manually tested resizing a few photos just to verify.
<details>
  <summary markdown="span">cat</summary>
  new (height)

![cat height new](https://user-images.githubusercontent.com/9317857/171272469-96e542c9-f26b-4192-ac8b-9ed8ef6e051d.png)
old (height)
![cat height old](https://user-images.githubusercontent.com/9317857/171272471-2cf10499-ee62-4f5d-a625-42e14cd6f5fb.png)
original
![cat](https://user-images.githubusercontent.com/9317857/171272472-10d5a368-63ac-46d7-9990-a410d93776ae.png)
</details>
<details>
  <summary markdown="span">4x3</summary>

new resize (height)
![4x3 height new](https://user-images.githubusercontent.com/9317857/171271796-e25fbee2-be1d-4069-b1c9-02442b122121.png)
old resize (height)
![4x3 height old](https://user-images.githubusercontent.com/9317857/171271797-7f40c64d-435f-447b-ab89-bf65caf1cb31.png)
new resize (width+height)
![4x3 new](https://user-images.githubusercontent.com/9317857/171271799-fce0a31b-dca1-44d3-8856-d4496f23a87c.png)
old resize (width+height)
![4x3 old](https://user-images.githubusercontent.com/9317857/171271801-6d17e571-386c-4ef3-9e7b-ada745394c7d.png)

</details>